tcp: Add support for AccECN
tx-buffer: Update () method to return number of bytes SACKed

Add AE flag in TCP Header for AccECN

Update ECN++ test for accepting 16 bit flag

AccEcn: Use uint16_t for TCP flags

Create state variable class for AccECN related variable

Add AccECN related state

Add AccECN related variables and there traces in TcpSocketBase

AccECN: Send AE with SYN

Decouple check CE in IP header from ForwardUp method using CheckEcnInIpv{4,6}

AccECN: Add DecodeAccEcnData () to decode AccECN related data from header

Handle ECE in case of Classic and EcnPp ECN

Add helpers to get and set AccEcn related flags

Update existing helper to support AccECN functionality

Adjust TcpSocketBase to use Enable AccEcn functionality using helpers

Update copyright and authors

Add AccECN test

Add AccECN example

Update tcp.rst for AccECN

Initial set of rebase changes

Rebased AccEcn

Minor edit

Minor edits

Review changes

+++ src/internet/test/tcp-ecnpp-test.cc
@@ -15,7 +15,8 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  *
- * Author: Wenying Dai <daiwenying927@gmail.com>
+ * Authors: Wenying Dai <dwy927@gmail.com>
+ *          Mohit P. Tahiliani <tahiliani.nitk@gmail.com>
  */
 
  #include "ns3/ipv4.h"
@@ -76,7 +77,7 @@ public:
   void SetTestCase (uint32_t testCase, SocketWho who);
 
 protected:
-  virtual void SendEmptyPacket (uint8_t flags);
+  virtual void SendEmptyPacket (uint16_t flags);
   virtual void PersistTimeout ();
   virtual Ptr<TcpSocketBase> Fork (void);
   void SetCE (Ptr<Packet> p);
@@ -114,7 +115,7 @@ TcpSocketCongestionRouter::Fork (void)
 }
 
 void
-TcpSocketCongestionRouter::SendEmptyPacket (uint8_t flags)
+TcpSocketCongestionRouter::SendEmptyPacket (uint16_t flags)
 {
   NS_LOG_FUNCTION (this << static_cast<uint32_t> (m_testcase) << static_cast<uint32_t> (flags));
   m_controlPacketSent++;
